+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 Version 11.1-SNAPSHOT
 -------------
 
+ADDED:
+
+- new factory method in class Zip4jExtensions for create a new ZipParameters object from the given parameters
 
 Version 11
 -------------